Added support for TestFixtures without explicit TestFixtureAttribute
