+++ OpenCover.UI.TestDiscoverer.TestResources/MSTest/RegularTestFixture.cs
@@ -1,5 +1,4 @@
 ﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
-using NUnit.Framework;
 
 namespace OpenCover.UI.TestDiscoverer.TestResources.MSTest
 {
