+++ OpenCover.UI.TestDiscoverer.TestResources/OpenCover.UI.TestDiscoverer.TestResources.csproj
@@ -52,6 +52,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="MSTest\RegularTestFixture.cs" />
+    <Compile Include="NUnit\TestFixtureWithoutExplicitTextFixtureAttribute.cs" />
     <Compile Include="NUnit\RegularTestFixture.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
